KEA-109 config sync

+++ backend/tests/helpers/strapi.js
@@ -3,15 +3,59 @@ const fs = require("fs");
 
 let instance;
 
+/**
+ * Setups strapi for futher testing
+ */
 async function setupStrapi() {
   if (!instance) {
     await Strapi().load();
-    instance = strapi;
-    
+    instance = strapi; // strapi is global now
+
     await instance.server.mount();
   }
   return instance;
 }
+
+/**
+ * Returns valid JWT token for authenticated
+ * @param {String | number} idOrEmail, either user id, or email
+ */
+const jwt = async (idOrEmail) =>
+  strapi.plugins["users-permissions"].services.jwt.issue({
+    [Number.isInteger(idOrEmail) ? "id" : "email"]: idOrEmail,
+  });
+
+/**
+ * Grants database `permissions` table that role can access an endpoint/controllers
+ *
+ * @param {int} roleID, 1 Autentihected, 2 Public, etc
+ * @param {string} value, in form or dot string eg `"permissions.users-permissions.controllers.auth.changepassword"`
+ * @param {boolean} enabled, default true
+ * @param {string} policy, default ''
+ */
+const grantPrivilege = async (
+  roleID = 1,
+  value,
+  enabled = true,
+  policy = ""
+) => {
+  const updateObj = value
+    .split(".")
+    .reduceRight((obj, next) => ({ [next]: obj }), { enabled, policy });
+
+  return await strapi.plugins[
+    "users-permissions"
+  ].services.userspermissions.updateRole(roleID, updateObj);
+};
+
+/** Updates database `permissions` that role can access an endpoint
+ * @see grantPrivilege
+ */
+
+const grantPrivileges = async (roleID = 1, values = []) => {
+  values.forEach((value) => grantPrivilege(roleID, value));
+};
+
 
 async function cleanupStrapi() {
   const dbSettings = strapi.config.get("database.connection");
@@ -25,10 +69,10 @@ async function cleanupStrapi() {
   //delete test database after all tests have completed
   if (dbSettings && dbSettings.connection && dbSettings.connection.filename) {
     const tmpDbFile = dbSettings.connection.filename;
-    if (fs.existsSync(tmpDbFile)) {
+    if (tmpDbFile == "test.db" && fs.existsSync(tmpDbFile)) {
       fs.unlinkSync(tmpDbFile);
     }
   }
 }
 
-module.exports = { setupStrapi, cleanupStrapi };
+module.exports = { setupStrapi, cleanupStrapi, grantPrivilege };
